Add 'id' to checkbox list
The 'aria-owns' attribute in the input of the searchable collection was
referencing an 'id' that didn't exist in the DOM structure.
This resulted in the association between the input field and checkbox list
being invalid, which was spotted in our accessibility audit and marked as an
error.

To resolve this, in this commit an 'id' attribute was added to the `div`
containing the checkboxes.

These changes enhance the accessibility of the searchable collection by creating a valid association between the containing input and the list of checkboxes. Now, assistive technologies would better understand the relationship between the input and the checkboxes, providing a better user experience for people using these technologies.