Combine first name and last name into a singular row called "name" on jobseeker profile form
Prior to this change we had accessibility issues stemming from the "Change first name" and "Change last name" links going to the same page. This change addresses the issue by combining both names under a "Name" row so that there is only one link to change name for a user on the jobseeker profile.