Rearrange footer elements
The accessibility assessment found an issue when navigating tabbing
through the footer elements.

The order elements are selected on the tabbing was going left to right
between sections instead of following the expected visual arrangement
from the top to the bottom.

Reorganising the footer sections using govuk grid elements allows us to
properly partition the footer elements to keep the same presentation
while achieving the tabbing expected behaviour.