Remove hidden text at start of link
Prior to this change, we had hidden text stating the number of jobs per link on the homepage, for example "View 2082 teacher jobs" where "View 2082" was visually hidden. The recently completed accessibility report found this to be an issue and advised that we should not have hidden text at the start of a link. This commit fixes this issue by removing the visually hidden text in question.