Move cookies banner to top of page
During the accessibility audit it was observed that while the cookie banner is in show, the user can still tab to the background content meaning that user interface components are obscured by this banner. It would be expected that, even though the cookie banner is on screen, the user can still see all content on the page when tabbed to. Moving the cookie banner to the top of the page should fix this issue.