Adapt Password input fields to new GovUK component
The Password field input coming with GovUK Frontend v5.3 has changed.
Now it includes a button to show/hide the password value from the input
for improving accessibility.

The new password field is larger than the previous one, and the helper
seems to be ignoring the "width" parameter.

This forced us to define a row & column width externally to achieve the
correct layout for the input fields.

We also made changes on the password input options to conform with
WCAG 2.2 accessibility standard.