Ensure tests always run on CI
All the tests jobs are "required" in Github in order to allow merging
the PR.

If a documentation change don't cause the tests to run, the required
tests checks are marked as expected/pendingm and the PR is unmergeable
unless a repo admin overrides the missing checks.

The problem now is that, if we allow admins to override the check
requirements to push into main branch, this also allows admins to
mistakenly push code directly into the main branch from their machines
(been there, done that).

Given the pros & cons. It is way simpler to always run the tests/checks
on github, regardless of code changes or not, and ensure the PRs are
clean for a merge without having to change settings or override
requirements.