Fix failing test due to date change
With the change of date, a Rails date helper using during the system
tests stopped returning the expected wording value.

To ensure we get the "about 1 year" in the wording. The date difference
must be (according to the helper docs): "1 yr <-> 1 yr, 3 months"

Setting the time gap between jobs is 1yr + 1 month, we ensure that the
date falls well within the range, instead of being on the edge between
two different wording outputs.