Skip MyNewTerm non-valid central trust jobs
When the central trust the vacancy is associated to is not a MAT, it
won't be listed in our system and no school will be found.

We are having issues with dozens of MyNewTerm vacancies being created
while not having any org associated to the due to this, causing
exceptions when listed/displayed in our system.

We are going to explicitly exclude those from being created.