Add mobile focus order for ayy1
On iOS voiceover, the elements on the page are being skippedaa on the profiles
question pages. The button appears to be the first element selected on load.
This fixes the issue by forcing the focus of the mobile page on the
first element of the page itself which is the skip to next section
button