Fix Sentry rule ignoring Safari parse JSON errors
There is a well known issue with Safari browsers parsing JSON
information. This triggers a JS browser error that doesn't affet the
user navigation, but causes tens of thousands of errors consuming our
Sentry quota.

We had a rule to ignore these so our quota doesn't get hit, but the
wording of the error slightly changed, causing the ignore rule to be
bypassed.