Fix links in publisher job application email alerts
The URL generated on the emails alerting publishers about job applications submitted
to their vacancies were returning 404/Not Found pages for the publishers upon clicking
on them.

The cause was the URLs were being generated using the friendly id/slug for the vacancies.

The publisher URLs do not use/accept the slug/friendly id, but require the real id for the
vacancy.