Fix DWP Find a Job vacancy description field
The "job_advert" field we were trying to parse is only present on
external vacancies.

For our internal vacancies, we use a combination of fields requested
during the job creation.

We need to reproduce the job description from TV combining those fields
in the parsed information we push to DWP Find a job description field.