Remove timetravel in tests
Due to the changes in the gem upgrade the time travel blocks do not work for out purpose in these tests. I believe this is because the noticed event is created inside the time travel block, but that creates noticed notifications in the background outside of the time travel block, so i have stubbed the created_at on the notifications to recreate these tests as best I can.