Remove flakey tests
These specs only test the timestamp method, and are very flaky due to the way the upgrade to the noticed gem works. Now the event is created when the deliver method is run and in turn this created notifications in the background, making these tests are flaky. I think this in combination with how little value these tests add make removing them a sensible idea.