Refactor of Find a Job export code
The principle behind this refactor is to use Composition instead of
Inheritance.

Now the Upload class is not a God object but a class with a single
responsibility: Upload the XML given contents into the DWP SFTP
server Inbox.

The orchrestration of the query-->xml-->upload is now managed by each
service class:
- PublishedAndUpdated
- ClosedEarly

The names of these classes have been adapted to match Teaching Vacancies
domain language.

These classes are not anaemic classes anymore.