Remove line breaks from jobs sent to find a job service
The stored HTML in DB used for the job descriptions contains carriage
returns that we do not want to display in the DWP Find a Job advert, as
they break sentences instead of respecting the HTML parsing structure
into plain text.

We are removing them prior to parse the HTML context into plain text for
the export.