Find a Job: Do not set explicit 30 days expiration
Find a job service only can have an expiration date for the vacancies
between 1 and 30 days from the publish/update date.

When we tried to push vacancies expicitly setting the expiration date to
"30 days from now" it gets rejected and the update doesn't get published
in the Find a Job service.

The rejection error is:
"Closing date must be between posting/editing date +1 day and posting
date +30 days"

If we leave the field empty, it will set the "30 days from now" date
when importing it on their side, as it is their default.

So we are reducing the maximum days to expicitly set to "29 days from
now", and if it is later than that, it will be defaulted by the service
to 30 days.