Do not mark applications without employment history gap validation as completed or imported
On April 3rd 2024 we introduced validation to ensure that the jobseeker could not submit a job application with an unexplained gap in their employment history. Applications prior to that may have unexplained gaps. This change ensures that if prefilling data from an application submitted before this date, we do not mark the employment history section as imported or completed and instead mark it as in progress to make sure the jobseeker has to account for any gaps in employment history.