Fix tests
When lacking personal details, the "full_name" for the Profile was
defaulting to "Jobseeker", than unfortunally matched the wording using
with the recent content changes.

So the assertions over the jobseeker name not being displayed were
giving false positives with unrelated page wording.

Adding real names in the profiles used for testing fixes it.