Repost vacancies every 30d to Find a Job service
The original attempt to keep our vacancies published for longer than 30
days ind Find a Job service failed, as they don't allow adverts to live
for longer than 30 days after they originally got posted.

The only way to keep our vacancies posted on Fins a Job for the whole
lifepan of the vacancy in our service is to keep reposting them every 31
days, just when the original/previous version of the advert reached its
days limit and is no longer posted.

They key concept in the new approach is "Vacancy versioning".

Vacancy versioning is based on how many days have passed since the
vacancy got published in TV.
- The original version of the vacancy will be published on Find a Job
  for 30 days after published on TV (and exported to Find a Job).
- 31 days after TV got published, it will be reposted under a new
  version.
- 62 days after TV got published, a further new version will be used to
  repost the vacancy.
- etc until the repost covers the period where the vacancy expires in
  our service.

To post different versioned adverts for the same vacancy, we will use
the vacancy uuid as a reference for the first advert.

After that, we will suffix the vacancy uuid with a "-1", "-2", "-3",...
Each corresponding with the particular "days since" version
corresponding with the current date.