Redirect user guides
We recently changed the paths to our user guides, however other services still link to the previous paths and we have recently sent out an email campaign with using the old, now not working, links. This change redirects to the appropriate pages. I have chosen to just make the changes in the routes path rather than add logic to the controller to avoid bloating it with code related to legacy links. I have hardcoded the redirects because we already have some constraints in the routes file which match jobseeker-guides and get-help-hiring.