Fix missing filter bug
Prior to this change, when a user filtered jobs by visa sponsorship availability, and then tried to set up a job alert using the existing filters, the visa sponsorship filter was not preselected on the job alert page. This was due to the visa_sponsorship_availability param not being permitted on the subscriptions controller. This change fixes this issue and adds this filter to the assertions on the job alert spec.