Setup Rails credentials with OneLogin keys
The OneLogin public/private pem keys are too big to be stored in AWS
Parameter Store.

We needed to find an alternative on how to secure this data.
Using the default Rails Credentials allows us to easily store the
encrypted keys in the project itself, while keeping the RAILS_MASTER_KEY
value in AWS Parameter store.