Fix OneLogin omniauth config breaking Publishers
Originally we were using Devise "omniauthable" module on the jobseekers.

This was breaking the system tests for the Publisher users (based on
Omniauth DFE Sign-in service).

We have found that "omniauthable" is not compatible with multiple models
using Omniauth. So we have to remove the "omniauthable" wrapper/config
over plain Omniauth for the Jobseeker, and configure it directly through
the Omniauth configuration, as done with DFE Sign-in Publisher users.

Devise documentation and instructions:
https://github.com/heartcombo/devise/wiki/OmniAuth-with-multiple-models