Sign-out Jobseekers through Govuk OneLogin
We had already built the endpoint and uri helpers to use the GovUK
OneLogin /logout endpoint, but the "Sign Out" link in our navigation bar
was still directly destroying the session through Devise.

The behaviour caused the jobseeker to be signed out from our service,
btu still signed-in in GovUK OneLogin. This caused the next sign-in
attempt to directly redirect back the jobseeker signed-in in our service
without having been prompted to any security user/password check by
OneLogin service.

We have fixed this by:
- Replacing the navigation Sign Out button to send the user to the
  /logout endpoint in GovUK One Login. This signs them out from
  OneLogin.
- Ensuring OneLogin redirects the user back to our internal session
  destroy endpoint. So the user session in our service is also signed
  out.

In order to achieve this, we had to enable a new route, to accept GET
request to the session destroying endpoint (defaults to DESTROY).