Fix Fusion ATS organisation mapping
Fusion was still using an old way to parse organisation info from the
feed. When the feed contained an incorrect School URN for a given Trust,
the resulting vacancy ended being not mapped to any School and breaking
the service when listed in the views.

Brought the new way to map organisations from the feed used in MyNewTerm
and other sources, that resolves this issue.