Fix login prompt after creating a subscription
The confirmation page after creating a job alert/subscription was
showing an old sign-in flow, requesting the password.

Fixed it by switching it to a One Login button/link and ensuring that
the users get redirected back to their job alerts dashboard (instead of
the Job applications dashboard) after they sign-in from this
confirmation page.