Remove registration link from forced login banner
The banner displaying "You need to log in or create an account..." when
an unsigned user attempts to save a vacancy or apply through a quick
apply vacancy contained a link to the old in-house user registration
flow.

Removing the link as we now manage registrations through GovUK One
Login.