Mitigate edge case when updating email from OneLogin
When a OneLogin user change their email in OneLogin service and we try
to update it in our system (EG: User realises they just signed up with
the wrong email and then change their email), we can still try to
automatically merge the Onelogin account with an existing Jobseeker.

This will happen automatically (like when signing up for first time) if
the OneLogin account had no saved data (applications, qualifications or
work history).

What we will do for this merge, as there will be 2 jobseekers in our DB,
is re-use the transfer account service in the background. Moving the
data from the old jobseeker to the OneLogin one, deleting the old
jobseeker and then updating the email for the OneLogin user.