Only validate models email if email changed
Since we introduced the new email validation, existing objects for
multiple models became invalid due to not passing the email validation.

This becomes an issue as affected users get unexpected errors when
trying to update those objects in different flows unrelated with
introducing/updating the email.

We are trying to mitigate this by only validate the email for the models
if the email field has been updated (new/changed value).

This would allow the object to still be valid when containing a legacy
email, while not allowing to create/edit the email with an invalid
value.