Remove nav for accessibility purposes
During an accessibility audit this page was found to have issues due to have two nav elements in close proximity. To fix this i have taken the applicable code from the sort component and hard coded it onto the page with the nav that wraps it removed. I did not remove the nav from the actual component because this component is used in multiple other places, none of which (as far as I know) are causing accessibility issues.