ATS imports skip vacancies without schools
There is a persistent issue when some ATS sources list vacancies with a
school URN that don't match any school in our system.

Sometimes those vacancies are being imported and, when listed, cause
errors for the jobseekers.

Instead of relying on the validation attempt prior to record the
vacancy, we are going to completely skip the vacancy from being even
build unless it is associated to an existing school.