Stub OneLogin RSA key in specs
Stubbing the RSA key, we isolate the test runs from the environment
having access to a real RSA key.

This was problematic on Dependabot-created PRs, as they had no access to
the Rails Credentials encrypted file.

We could set some dummy private key in the test environment values, but
I suspect it would be raised by security tools as "exposing a secret"
commited in git.

So I opted for stubbing the RSA key when needed..