Fix Alert Email jobs memory load
We're experience serious performance issues and pods restarting during
the Daily/Weekly alert email job runs.

At the moment we have over 100k subscriptions for daily alerts, and over
30k subscriptions for weekly alerts.

The existing code, loads all the subscriptions in memory prior to
iterate over them.

We are changing this to be fetch them in batches, aiming to improve the
memory load and job performance.