Fix Ventrus school mapping
There is a bug in the existing code when no school URN comes from
Ventrus feed. Our system assings the vacancy to the first Organisation
found with "nil" urn value.

So the Ventrus vacancy ends assigned to a random MAT/SchoolGroup (that
has no URN value).

To fix it, we are:
- Ensuring that the correct trust ID must be provided by Ventrus feed.
- Only assigning the vacancy to a particular school if the URN is
  provided and belongs to a school that is part of Ventrus.
- When no/incorrect school urn is provided, it assigns it to Ventrust
  trust itself, not to a particular school.