Move API from "api/v2" to "ats-api/v1"
The original API in the service is an API that the public can consume to
get information about the posted vacancies.

The new API is going to be private and only accessible to the ATS so
they can Publish their vacancies into our system.

To avoid confusion, extracted the new API definition under the following
path in our project:

"app/controller/publishers/ats_api/v1/...."

So developers can identify the scope of this API.

To avoid a long convoluted API endpoint for our clients, we will expose
it under "/ats-api/v1" path.