Do not use cached swagger docs when config changed
When building the documentation for swagger, it keeps the old
documentation as a "cache hit" even when the swagger schemas have been
changed, or if the controller implementation (used for generating
automatic examples on the docs) changed.

Adding both the swagger helper and api controller files to the hashed
files used for generating the cache key should resolve this, as when any
of those files is touched, the swagger cache will become invalid and a
new cached version will be generated.