Only generate swagger docs if needed
The generation of swagger documentation seemed to be happening every
time, and then being overridden by the cached version.

We are changing them to first check if the cached version can be
used/restored, and only generating the docs if the attempt to find and
restore a cached version failed.

The "cache" action has a post step to save the cache. So the documents
generated in the "swagger" path will be cached afterwards.