Fix "data" arg format for all analytics events
dfe-analytics v1.13 introduced a unanounced (and seems unexpected)
breaking change on how the "data" argument must be formatted for the
events.

Since then, it seems we adapted our calls in some instances when we
wanted/needed to use the new 'hidden_data' option.

The problem is, as much as in their documentation state that you can
call a custom event with `.with_data(some: "info")`, in reality all the
events instantiated like that get their data dropped unless are wrapped
within a `data:` key: `.with_data(data: { some: "info" })`