Precompute Centroid for Location Polygons
The centroid is used for calculating the distances on the vacancy
searches.

The centroid for a polygon area doesn't change unless the area changes
too.

As we dinamically calculate it for every location polygon area
against every vacancy during the searches, if we pre-compute it we
save the DB to execute again the centroid calculation on every search.