Set option for faster geography evaluation
By setting the "use_spheroid = false" flag in the ST_DWithin and
ST_Distance PostGis over geography data, the operation benefits by a
faster spherical calculation rather that the default computing on the
spheroid determined by the SRID.

The downside is a slightly less accurate computation. But seems a good
tradeoff since we're experiencing performance issues on the distance
calculations and our distances are not that big to make a huge
difference on the seach results.

Comment on why we repeat query for ordering

Ideally we would use the 'distance' alias on the ordering, but when the
location query gets combined with a 'pluck(:id)' (as it happens) call,
the generated SQL query ignores the first 'select ... AS distance' bit
of the query, and when 'distance' is called in the ordering, triggers an
exception.

For the time being, lets keep trusting the PostgreSQL query planner that
doesn't recalculate the repeated distance calculation (when already done
in the same query).