Remove robots/crawlers exclusion of sign-in page
The sign-in page now just points to external Oauth integrations for both
Jobseekers and Hiring Staff.

Some search engines are providing a preview box of this (and others) link
when teaching vacancies is returned. The link context shows some
unhelpful message: "We would love to show you a description here but the
site won't allow us".

Given we don't find downsides on allowing crawlers to visit this page,
we are removing the exclusion so the search engines can list it.