Extend cached period of Google Geocoding results (#7514)
We're attempting to cut costs on the Google Geocoding bill.

Currently, we query the API around 4.26k times per day on average. The
volume of location searches received by the server exceeds 70k searches
per day.

As the geocoding information within the UK is very stable data, keeping
the results cached for only 30 days seems quite a conservative approach.

By increasing it to 180 days we aim to reduce the amount of times the
same location searches get re-queried to the API.