Simplify location polygon area for performance
The original ONS location polygons contains tens of thousands of
vertices (125k for Cornwall).

By applying a simplification with some tolerance we can reduce this
massively while keeping a quite precise resulting area that retains the
same shape as the original (with some minor innacuracies of course).

This should be good enough for what we use the polygons for, and should
help with the big performance issues we're dealing with when calling
ST_BUFFER method over some of the polygons.