Fix bug which causes job preferences page to crash if a JobPreference… (#7526)
This PR fixes [this
bug](https://teaching-vacancies.sentry.io/issues/5991105845/?environment=production&project=6212514&query=is%3Aunresolved%20issue.priority%3A%5Bhigh%2C%20medium%5D%20NoMethodError&referrer=issue-stream&statsPeriod=30d&stream_index=0)
which has been causing us issues since October.

We create an instance of JobseekerProfile when the user visits the
jobseeker profile page. We do this using
[JobseekerProfile#prepare](https://github.com/DFE-Digital/teaching-vacancies/blob/main/app/models/concerns/profile_section.rb)
which should in turn create an associated instance of JobPreferences and
PersonalDetails. It seems like this sometimes fails which causes the
above error when users try to visit the job preferences page and job
preferences do not exist. This change fixes the issue.

Note: I did not need to do the same for personal details as we already
build an instance of PersonDetails if one does not exist (as can be seen
[here](https://github.com/DFE-Digital/teaching-vacancies/blob/dfd4ba12798acd25adc46b9d2c13b5b82acf9824/app/controllers/jobseekers/profiles/personal_details_controller.rb#L22)),
but for some reason we were not doing the same for job preferences.