Simplify location polygon area for performance (#7527)
## Trello card URL
-
https://trello.com/c/XjF2cQcC/1602-fix-performance-issues-with-new-location-polygon-areas
## Changes in this PR:
- Simplify our location polygon areas to have way fewer points/vertices
than the originals from ONS.
- Makes invalid polygons valid before storing them.

These changes tackle a massive performance hit that cause DB restarts
when buffering some polygon areas for location searches.

### Extended
The original ONS location polygons contain tens of thousands of vertices
(125k for Cornwall).

By applying a simplification with some tolerance we can reduce this
massively while keeping a precise-enough resulting area that retains the
same shape as the original (with some minor inaccuracies of course).

This should be good enough for what we use the polygons for and help
with the big performance issues we're dealing with when calling the
`ST_BUFFER` method over some of the polygons.

After the simplification, the resulting polygon is ensured to be valid
by applying the `ST_MakeValid` operation over the simplified output.

## Screenshots of UI changes:

### Superposed original Cornwall area with 125k vertices and our
simplified area with 2.5k vertices after applying a `0.001` tolerance
simplification.


![image](https://github.com/user-attachments/assets/a378b793-6bf0-4e67-8fc0-ae52f75e6ab8)

#### Zoomed in to notice the actual differences 

![image](https://github.com/user-attachments/assets/da479b52-078c-4c46-82f6-45788b078ee7)

### Superposed original Warwickshire area with 18k points vs our
simplified area with 565 points after applying same `0.001` tolerance
simplification


![image](https://github.com/user-attachments/assets/531936f3-1739-40fe-90ea-4b165f892ae5)

#### Zoomed in to see the differences


![image](https://github.com/user-attachments/assets/e8b6274b-6f6c-480d-bcc2-99b5b9436f89)