Add empty model class to allow old data migration
If the data migration runs against the up to date todays model class,
the attributes/methods may differ from the ones existing/needed when the
old migration was created.

This causes issues when running old migrations as exceptions arise due
to the missing method/attribute.

By creating a minimal empty model class we remove the dependency on the
real model.