Flexi work question (#7533)
This change adds a radio button to the vacancy form, which asks hiring staff
if they want to add flexible working details to their application form.
The reason for this is that a number of hiring staff were adding N/A or
similar to the flexible working details section rather than leaving it
blank as expected.