Remove external TRI setting in favour of existing form logic
There was a bit of unreachable code in the Jobseeker Professional Status
Form.

This code was setting the TRI (teacher reference information) on the
form intialisation from the jobseeker profile if provided and the form
has no explicit TRI values provided.

The problem is that the from wasn't receiving the jobseeker profile when
initialised. So this code was unreachable (and untested, as the code
coverage tool has proven quite useful to identify this).

The same logic to "use the TRI stored in the profile to set the form
attributes value if not set" was also defined and used at the controller
level.

I've opted out to remove this second option in favour of form object
approach, as when defined in the form the specific TRI logic is more
contained within its domain instead of in a shared controller, and also
allows for more extensive unit testing.