Use escape path for back_link from personal details name page (#7553)
We recently had a user report a bug. The user could not turn on their
profile because they had not completed their personal details. The name
page crashed when they tried to enter their name in the jobseeker
profile.

### Why did this happen?

The user's jobseeker profile had some personal details filled in but the
name section was blank. In completed_steps it had work and phone_number
sections marked as complete but not the name section. The name is the
first step in the personal details flow.

The "Step not completed" error was raised by the `previous_step` defined
in lib/multistep/form.rb. This error seems to be raised to ensure that
multistep forms cannot be filled in out of order. As name was not
completed but later steps in the personal details flow were completed it
raised this error.

The `previous_step` method was called because of the back link on the
name page of the personal details form.

### How did this happen?

I think the user's jobseeker profile got into this state due to the way
that we populate a new jobseeker profile for a user. If the user has a
job application, when they first visit the jobseeker profile page, we
populate their profile with information from their previous application
regardless of whether that application has been submitted or not.

In this case the user had both a draft and a submitted job application,
I think that the user's jobseeker profile got populated based on the
draft application which was missing some information.

### How to fix this?

I have tried to fix this both for other jobseekers who may already have
their data in a similar state and to ensure other jobseekers do not get
their data into a similar state.

1. Using escape_path for the back link from the name page of the personal details
flow. Removing this back link stops previous_step being called and stops the
error being raised for users who find themselves in a similar position
to the user who raised the issue.

2. Only populate new jobseeker profiles with data from job applications
if they are not in a draft state. This hopefully should ensure that the
data from the job application is completed and valid and will not create
a jobseeker profile that is missing important data.