TRN form tweaks: Remove duplicated code, reorganise validations and reduce duplicated validation messages. (#7560)
## Trello card URL

-
https://trello.com/c/2LLFZmsR/1632-remove-repeated-validation-errors-on-professional-status-form-teacher-reference-number-attributes

## Changes in this PR:

There are 3 main changes on the job application professional status form
submission:

1. Code to assign the TRN attributes from the jobseeker profile was
defined in 2 different places: The form itself and the controller.
    Only the controller code was reachable. 
This PR removes the but from controller logic and moves it to the form,
as it aims to keep as much of the particular form logic within the form
itself (especially when the controller is shared across the whole job
application flow) and it allows deeper unit testing at form object model
level.

2. Refactor validations grouping them by shared conditions, and add
extensive validation testing.
This was mainly done to be able to identify and fix any redundancy (next
point), but also made the form validations easier to glance through and
more manageable.

2. Remove duplications on professional status form error messages.
The combination of attribute validations, presence, format, inclusion...
caused situations where the same error message was displayed for 2
different fields, or the same field has the same error message listed
twice.
We have simplified the validations over TRN fields so they don't step on
each other.

## Screenshots of UI changes:

### Before


![Screenshot_20250226_204224](https://github.com/user-attachments/assets/5ee819a5-7994-4612-aca7-682af6348c68)


![Screenshot_20250226_222455](https://github.com/user-attachments/assets/f025c687-f3b3-4a49-89f4-f37d5cea91a1)

### After


![image](https://github.com/user-attachments/assets/496eb42e-7b3d-441b-a04c-771e3372002a)


![image](https://github.com/user-attachments/assets/359d3643-3b18-47d1-a25a-63b50ebccb7b)

#### Only if selecting "Yes" on "Do you have a teacher reference number
(TRN)?" but not entering the number

![image](https://github.com/user-attachments/assets/f8f5b9b7-c054-4968-9fba-2481521610a1)