Spike: move TRN input  to different QTS options
Aiming to simplify validations and logic over the TRN requirement on the
job application professional status step.

Instead of explicitly asking if the user has a TRN independently from
the QTS, we embed the TRN field in each of the possible radio button
options on the QTS question.

Then it only is required if the QTS answer is "Yes".