Fix campaign pages subjects (#7588)
## Changes in this PR:
**Fix campaign pages subjects** 

The subjects listed in the campaign pages weren't matching our subjects
list in the filters, what caused the desired subjects not being
automatically applied per campaign.

**Add configuration tests for campaign filters**
Ensure each one of the values present in the campaigns configurations
matches an:
- Existing banner image.
- Existing teaching job role.
- Existing subject.

So, if the wrong value is entered when adding/updating a campaign, it
will be caught by our tests.

## Screenshots of UI changes:

### Ensuring campaign banner image is valid
![Screenshot From 2025-03-06
13-03-04](https://github.com/user-attachments/assets/b328f95c-bee8-4e4a-9c3f-ec0d7e052795)

### Ensuring campaign teaching job role is valid
![Screenshot From 2025-03-06
13-24-13](https://github.com/user-attachments/assets/81c411e3-363f-4fc3-8dea-1d2e6b119424)

### Ensuring campaign subject is valid
![Screenshot From 2025-03-06
13-32-37](https://github.com/user-attachments/assets/dfa34b59-7a53-40be-9a2e-2423e4ad3c36)