Remove skills_and_experience from API specification
The job description comes in the job_advert field.

The skills and experience is used in manual vacancies and if present,
the information is shown instead of the job_advert.

For external vacancies we do not expect them to provide the different
bits for a job description text, so everything comes together in the
job_advert.