Fix ECT Status parsing
We provide a boolean interface "ect_suitable: true/false" to the
integrations. While we store it as an enumerable on "ect_status".

Fixing this.

Also ensuring that it defaults to "false" unless explicitly set as
"true", and that the original value is kept if the optional field is not
provided in further vacancy update requests.