Extract API Schemas to shared swagger helper conf
Multiple API schemas for the API were defined within the API tests and
documentation file.

The Create and Update schemas were duplicated. And everytime we needed
to change the schema, this needed to be done in multiple definitions and
files.

This also made the tests very hard to read/follow, as the schema setups
were very long and made the actual test code to be hard to find.

Normalised this by:
- Defining the schemas in a single file (swagger_helper).
- Referencing the schemas from the test file.
- Remove fields requirement in response schemas.