Fix starts_on parsing on creation/update
On the vacancies created through ATS imports/API, there is some data
parsing needed to be done to allow both exact start date or text as
input in the "starts_on" field.

Luckly we already have a class that dealt with that parsing for the
import vacancies legacy process. So re-using that in the API is quite
straight.