Fix: Vacancy responses contain all the fields
Normalise as a good practice that all our responses containing vacancy
objects always contain the same fields for the vacancy (present or not).

This ensures a stable interface the API clients can rely on when
querying their existing vacancies.

Add tests to ensure the values are what we expect (as for some fields
there is some parsing involved) and that all the fields presence is
required by the schema definition for the responses.