Fix org fetching exception
Do not assume that if the org slug/id doesn't match the current
user organisation, it means the current org is a SchoolGroup and can be
queried as so.

Introduce explicit condition to only query it as a SchoolGroup if it is
one.

If neither options on the retrieval match, return a controlled 404
error.