Fix org fetching exception (#7655)
## Trello card URL

Fixes:
https://teaching-vacancies.sentry.io/issues/6110280649/events/ccf31849dc2041849ce79f6d19c9ae1e/

## Changes in this PR:
Do not assume that if the org slug/id doesn't match the current user
organisation, it means the current org is a SchoolGroup and can be
queried as so.

Introduce an explicit condition to only query it as a SchoolGroup if it
is one.

If neither options on the retrieval match, return a controlled 404
error.

## Screenshots of UI changes:

Accessing the wrong id for the academy:
`http://localhost:3000/publishers/organisations/bexleyheath-acad`
### Before


![image](https://github.com/user-attachments/assets/e7638a38-8a80-4552-a32b-381d130fb075)

### After

![Screenshot From 2025-03-27
12-14-56](https://github.com/user-attachments/assets/dcd1c561-1ac5-4ca7-be98-e548db448a31)

## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>